Blew away the data and am in the process of fixing areas of code where things don't working intially.

+++ GreenBushIEP/Models/ManageViewModels.cs
@@ -114,7 +114,7 @@ namespace GreenBushIEP.Models
         public string ImageURL { get; set; }
         public Nullable<bool> Agreement { get; set; }
         public Nullable<bool> Archive { get; set; }
-        public int? KidsID { get; set; }
+        public long? KidsID { get; set; }
         public DateTime DateOfBirth { get; set; }
         public string Status { get; set; }
         public string BuildingName { get; set; }
